+++ widgetastic_manageiq.py
@@ -20,7 +20,8 @@ from widgetastic.widget import (
 from widgetastic.utils import ParametrizedLocator
 from widgetastic.xpath import quote
 from widgetastic_patternfly import (
-    Accordion as PFAccordion, CandidateNotFound, BootstrapTreeview, Button, Input, BootstrapSelect)
+    Accordion as PFAccordion, CandidateNotFound, BootstrapTreeview, Button, Input, BootstrapSelect,
+    Dropdown)
 from cached_property import cached_property
 from utils.log import logger
 
@@ -898,7 +899,6 @@ class Stepper(Widget):
     def __locator__(self):
         return self._stepper_control
 
-    @property
     def read(self):
         return int(self.value_field.read())
 
@@ -920,6 +920,9 @@ class Stepper(Widget):
         steps = abs(steps)
         for step in range(steps):
             operation()
+
+    def fill(self, value):
+        self.set_value(value)
 
 
 class RadioButton(Button):
@@ -960,10 +963,164 @@ class RadioGroup(Widget):
             btn = [btn for btn in self.buttons if self.browser.text(btn) == button_name][-1]
             btn.click()
 
+    def read(self):
+        return self.selected
+
+    def fill(self, name):
+        self.select(name)
+
 
 class BreadCrumb(Widget):
     """ CFME BreadCrumb navigation control
 
+        locator should be [@class='breadcrumb']
+    """
+    def __init__(self, parent, locator, logger=None):
+        super(BreadCrumb, self).__init__(parent=parent, logger=logger)
+        self.locator = locator
+        self._locations = self.browser.elements('.//li', parent=self.locator)
+
+    def __locator__(self):
+        return self.locator
+
+    @property
+    def locations(self):
+        return [self.browser.text(loc) for loc in self._locations]
+
+    @property
+    def active_location(self):
+        br = self.browser
+        return [br.text(loc) for loc in self._locations if 'active' in br.classes(loc)][-1]
+
+    def click_location(self, name):
+        br = self.browser
+        location = [loc for loc in self._locations if br.text(loc) == name][-1]
+        br.click(location)
+
+
+class ToolBarViewSelector(View):
+    """
+    represents toolbar's view selector control.
+    """
+    list_view = Button(title='List View')
+    grid_view = Button(title='Grid View')
+    tile_view = Button(title='Tile View')
+
+    def __locator__(self):
+        return './/div[contains(@class, "toolbar-pf-view-selector")]'
+
+    def select(self, item):
+        if item == self.list_view.title:
+            self.list_view.click()
+        elif item == 'Grid View':
+            self.grid_view.click()
+        elif item == 'Tile View':
+            self.grid_view.click()
+        else:
+            raise ValueError('Incorrect value passed')
+
+    @property
+    def selected(self):
+        """
+        goes thru buttons and returns the title of active button
+        Returns: currently selected view
+
+        """
+        return [btn.title for btn in (self.list_view, self.grid_view, self.tile_view)
+                if btn.active][-1]
+
+
+class ProviderToolBar(View):
+    """
+    represents provider toolbar and its controls
+    """
+    configuration = Dropdown(text='Configuration')
+    policy = Dropdown(text='Policy')
+    authentication = Dropdown(text='Authentication')
+    download = Dropdown(text='Download')
+
+    @View.nested
+    class view_selector(ToolBarViewSelector):  # NOQA
+        pass
+
+
+class DetailsProviderToolBar(View):
+    """
+    represents provider toolbar and its controls
+    """
+    monitoring = Dropdown(text='Monitoring')
+    configuration = Dropdown(text='Configuration')
+    reload = Button(title='Reload Current Display')
+    policy = Dropdown(text='Policy')
+    authentication = Dropdown(text='Authentication')
+
+    @View.nested
+    class view_selector(ToolBarViewSelector):  # NOQA
+        # todo: there should be another ViewSelector. to add it later
+        pass
+
+
+class Items(View):
+    """
+    should represent the view with different items like providers
     """
 
+    @View.nested
+    class search(object):  # NOQA
+        pass
+
+
+class BaseSideBar(View):
+    """
+    represents left side bar. it usually contains navigation, filters, etc
+    """
     pass
+
+
+class TimelinesFilter(View):
+    def __init__(self, parent, filter_type, logger=None):
+        super(TimelinesFilter, self).__init__(parent=parent, logger=logger)
+        if filter_type in ('Management Events', 'Policy Events'):
+            self._filter_type = filter_type
+        else:
+            raise ValueError('incorrect filter type is passed')
+
+    # common
+    event_type = BootstrapSelect(id='tl_show')
+    event_category = BootstrapSelect(id='tl_category_management')
+    time_period = Stepper()
+    time_range = BootstrapSelect(id='tl_range')
+    time_position = BootstrapSelect(id='tl_timepivot')
+    date_picker = Calendar()
+    apply = Button("Apply")
+    # management controls
+    detailed_events = Checkbox(name='showDetailedEvents')
+    # policy controls
+    event_category = BootstrapSelect(id='tl_category_policy')
+    event_status = RadioGroup()
+
+
+class TimelinesChart(View):
+    # todo: to add widgets for all controls
+    # currently only events collection is available
+    pass
+
+
+class Timelines(View):
+    """
+    represents Timelines page
+    """
+
+    @View.nested
+    class sidebar(BaseSideBar):  # NOQA
+        pass
+
+    breadcrumb = BreadCrumb()
+
+    @View.nested
+    class filter(TimelinesFilter):  # NOQA
+        pass
+
+    @View.nested
+    class chart(TimelinesChart):  # NOQA
+        pass
